Extract an interface for the FilterFromDialog
This allows multiple editable filter implementations.

Issue: #162
