+++ src/name/mlopatkin/andlogview/ui/filterdialog/FilterFromDialogImpl.java
@@ -1,0 +1,71 @@
+/*
+ * Copyright 2024 the Andlogview authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package name.mlopatkin.andlogview.ui.filterdialog;
+
+import name.mlopatkin.andlogview.filters.ColoringFilter;
+import name.mlopatkin.andlogview.filters.FilteringMode;
+import name.mlopatkin.andlogview.logmodel.LogRecord;
+
+import org.checkerframework.checker.nullness.qual.Nullable;
+
+import java.awt.Color;
+
+public class FilterFromDialogImpl implements FilterFromDialog, ColoringFilter {
+    private final boolean enabled;
+    private final FilterFromDialogData data;
+
+    FilterFromDialogImpl(boolean enabled, FilterFromDialogData data) {
+        this.enabled = enabled;
+        this.data = data;
+    }
+
+    @Override
+    public boolean test(LogRecord input) {
+        return data.test(input);
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    @Override
+    public FilteringMode getMode() {
+        return data.getMode();
+    }
+
+    @Override
+    public @Nullable Color getHighlightColor() {
+        return data.getHighlightColor();
+    }
+
+    @Override
+    public FilterFromDialogImpl enabled() {
+        return enabled ? this : new FilterFromDialogImpl(true, data);
+    }
+
+    @Override
+    public FilterFromDialogImpl disabled() {
+        return enabled ? new FilterFromDialogImpl(false, data) : this;
+    }
+
+    @Override
+    public FilterFromDialogData getData() {
+        // TODO(mlopatkin) defensive copy?
+        return data;
+    }
+}
